requirements.txt
--index-url https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/cpu

# =========================
# Core ML / Embeddings
# =========================
torch==2.6.0+cpu
sentence-transformers==3.3.1
numpy==1.26.4
scipy==1.11.4

# =========================
# FastAPI Stack
# =========================
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1

# Pydantic v2 ecosystem
pydantic==2.6.4
pydantic-settings==2.2.1

# =========================
# LangChain / LangGraph
# =========================
langchain==0.1.16
langchain-community==0.0.34
langchain-openai==0.1.3
langchain-aws==0.1.6
langgraph==0.0.38

langsmith==0.1.31

# =========================
# Vector Stores / Search
# =========================
pinecone-client==3.2.2
qdrant-client==1.8.2
opensearch-py==2.4.2

# =========================
# Memory / Databases / Cache
# =========================
redis==5.0.3
psycopg2-binary==2.9.9
asyncpg==0.29.0
gptcache==0.1.43

# =========================
# AWS
# =========================
boto3>=1.34.69
aioboto3>=12.3.0

# =========================
# WebSockets
# =========================
websockets==12.0
python-socketio==5.11.1

# =========================
# Monitoring / Tracing / Eval
# =========================
ragas==0.1.7
prometheus-client==0.20.0
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0

# =========================
# HTTP / Async Utilities
# =========================
httpx==0.27.0
aiohttp==3.9.3

# =========================
# Tokenization / Data
# =========================
pandas==2.2.1
tiktoken==0.6.0

# =========================
# Testing
# =========================
pytest==8.1.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# =========================
# Dev / Linting / Typing
# =========================
black==24.3.0
flake8==7.0.0
mypy==1.9.0
